code/EduTutor-AI/requirements.txt
streamlit
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
authlib==1.3.0
requests==2.31.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
langchain==0.2.0
langchain-ibm==0.1.3
pinecone-client==4.0.0
streamlit-authenticator
PyYAML 